+++ contract/build.gradle
@@ -2,18 +2,13 @@ apply plugin: 'kotlin'
 apply plugin: 'kotlin-jpa'
 apply plugin: 'kotlin-allopen'
 apply from: '../deterministic.gradle'
-apply plugin: 'net.corda.plugins.cordapp'
-apply plugin: 'maven-publish'
 apply plugin: 'idea'
-
-cordapp {
-    info {
-        vendor = 'Cordapp Supplier'
-    }
-}
 
 dependencies {
     cordaCompile "net.corda:corda-serialization-deterministic:$corda_release_version"
+    cordaCompile("net.corda:corda-jackson:$corda_release_version") {
+        exclude group: 'net.corda'
+    }
     cordaCompile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
     cordaCompile "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
     compileOnly "org.hibernate:hibernate-core:$hibernate_version"
@@ -29,19 +24,8 @@ allOpen {
     )
 }
 
-task install(dependsOn: publishToMavenLocal)
-
 jar {
     baseName "${rootProject.name}-contract"
-}
-
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            from components.java
-            artifactId jar.baseName
-        }
-    }
 }
 
 idea {
