+++ contract/src/main/kotlin/com/example/goody/contracts/Candy.kt
@@ -1,0 +1,9 @@
+package com.example.goody.contracts
+
+import com.fasterxml.jackson.annotation.JsonProperty
+import net.corda.core.serialization.CordaSerializable
+
+@CordaSerializable
+data class Candy(@JsonProperty("type") val type: String) {
+    override fun toString(): String = type
+}
