--- contract/src/main/kotlin/org/testing/goody/contracts/Candy.kt
@@ -1,6 +1,0 @@
-package org.testing.goody.contracts
-
-import net.corda.core.serialization.CordaSerializable
-
-@CordaSerializable
-data class Candy(val type: String)
