+++ flow/build.gradle
@@ -1,17 +1,8 @@
 apply plugin: 'kotlin'
-apply plugin: 'net.corda.plugins.cordapp'
 apply plugin: 'net.corda.plugins.quasar-utils'
-apply plugin: 'maven-publish'
-apply plugin: 'idea'
-
-cordapp {
-    info {
-        vendor = 'Cordapp Supplier'
-    }
-}
 
 dependencies {
-    implementation(project(':contract')) {
+    cordapp(project(':contract')) {
         transitive = false
     }
     cordaCompile "net.corda:corda-node-api:$corda_release_version"
@@ -22,17 +13,6 @@ dependencies {
     testImplementation "junit:junit:$junit_version"
 }
 
-task install(dependsOn: publishToMavenLocal)
-
 jar {
     baseName "${rootProject.name}-flow"
-}
-
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            from components.java
-            artifactId jar.baseName
-        }
-    }
 }
