+++ flow/src/main/kotlin/com/example/goody/flows/GoodyException.kt
@@ -1,0 +1,5 @@
+package com.example.goody.flows
+
+import net.corda.core.flows.FlowException
+
+open class GoodyException(message: String, cause: Throwable?) : FlowException(message, cause)
