+++ flow/src/main/kotlin/com/example/goody/flows/InsufficientGoodiesException.kt
@@ -1,0 +1,5 @@
+package com.example.goody.flows
+
+class InsufficientGoodiesException(message: String, cause: Throwable?) : GoodyException(message, cause) {
+    constructor(message: String) : this(message, null)
+}
