--- flow/src/main/kotlin/org/testing/goody/flows/AbstractGoodyFlow.kt
@@ -1,20 +1,0 @@
-package org.testing.goody.flows
-
-import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.flows.FinalityFlow
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.NotaryException
-import net.corda.core.identity.Party
-import net.corda.core.transactions.SignedTransaction
-import net.corda.core.utilities.ProgressTracker
-
-abstract class AbstractGoodyFlow(override val progressTracker: ProgressTracker) : FlowLogic<SignedTransaction>() {
-    @Suspendable
-    protected fun finaliseTx(tx: SignedTransaction, extraParticipants: Set<Party>, message: String): SignedTransaction {
-        try {
-            return subFlow(FinalityFlow(tx, extraParticipants))
-        } catch (e: NotaryException) {
-            throw GoodyException(message, e)
-        }
-    }
-}
