--- flow/src/main/kotlin/org/testing/goody/flows/GoodyException.kt
@@ -1,5 +1,0 @@
-package org.testing.goody.flows
-
-import net.corda.core.flows.FlowException
-
-open class GoodyException(message: String, cause: Throwable?) : FlowException(message, cause)
