--- flow/src/main/kotlin/org/testing/goody/flows/InsufficientGoodiesException.kt
@@ -1,5 +1,0 @@
-package org.testing.goody.flows
-
-class InsufficientGoodiesException(message: String, cause: Throwable?) : GoodyException(message, cause) {
-    constructor(message: String) : this(message, null)
-}
