+++ settings.gradle
@@ -1,3 +1,4 @@
 rootProject.name = 'cordapp-example-deterministic'
 include 'contract'
 include 'flow'
+include 'web'
