+++ web/src/main/kotlin/com/example/goody/plugin/GoodyPlugin.kt
@@ -1,0 +1,9 @@
+package com.example.goody.plugin
+
+import com.example.goody.api.GoodyApi
+import net.corda.webserver.services.WebServerPluginRegistry
+import java.util.function.Function
+
+class GoodyPlugin : WebServerPluginRegistry {
+    override val webApis = listOf(Function(::GoodyApi))
+}
