+++ web/src/main/resources/META-INF/services/net.corda.webserver.services.WebServerPluginRegistry
@@ -1,0 +1,1 @@
+com.example.goody.plugin.GoodyPlugin
